Upload image

+++ client/src/pages/BlogPostPage.jsx
@@ -2,12 +2,12 @@ import React, { useEffect, useState } from "react";
 import Card from "@mui/material/Card";
 import CardActions from "@mui/material/CardActions";
 import CardContent from "@mui/material/CardContent";
-import CardMedia from "@mui/material/CardMedia";
 import Button from "@mui/material/Button";
 import Typography from "@mui/material/Typography";
 import { useParams, useNavigate } from "react-router-dom";
-// import EditPost from "./EditPost/EditPost";
 import { useDispatch, useSelector } from "react-redux";
+import { openEdit } from "../store/editPost";
+import EditPost from "./EditPost/EditPost";
 
 export default function BlogPostPage() {
   const [post, setPost] = useState({});
@@ -17,22 +17,26 @@ export default function BlogPostPage() {
   const navigate = useNavigate();
 
   function editHandler() {
-    // dispatch(openEdit());
+    dispatch(openEdit());
   }
 
   async function deleteHandler() {
-    let response = await fetch(`http://localhost:4000/deletePost`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ id: +params.blogId }),
-    });
-    if (!response.ok) {
-      alert(response.message);
-      return;
+    // eslint-disable-next-line no-restricted-globals
+    let confirmed = confirm("Are you sure?");
+    if (confirmed) {
+      let response = await fetch(`http://localhost:4000/deletePost`, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({ id: +params.blogId }),
+      });
+      if (!response.ok) {
+        alert(response.message);
+        return;
+      }
+      navigate("/blogs");
     }
-    navigate("/blogs");
   }
 
   useEffect(() => {
@@ -66,6 +70,14 @@ export default function BlogPostPage() {
 
   return (
     <>
+      {isEdit === "show" && (
+        <EditPost
+          id={post.id}
+          title={post.title}
+          content={post.content}
+          imageData={post.imageData}
+        />
+      )}
       {/* {isEdit === "show" && (
         <EditPost
           id={post.id}
@@ -83,32 +95,43 @@ export default function BlogPostPage() {
             justifyContent: "space-between",
           }}
         >
-          <>
-            <CardMedia
+          <div style={{ padding: "1rem" }}>
+            {/* <CardMedia
               component="img"
               alt={post.title}
               src={post.imageData}
               title={post.title}
+            /> */}
+            <div
+              style={{
+                backgroundImage: `url(${post.imageData})`,
+                backgroundPosition: "center",
+                backgroundRepeat: "no-repeat",
+                backgroundSize: "cover",
+                height: "100vw",
+              }}
             />
-            <CardContent>
-              <Typography gutterBottom variant="h5" component="div">
-                {post.title}
-              </Typography>
-            </CardContent>
-            <CardContent>
-              <Typography gutterBottom variant="p">
-                {post.content}
-              </Typography>
-            </CardContent>
-            <CardActions sx={{ height: 50 }}>
-              <Button size="small" onClick={editHandler}>
-                Edit
-              </Button>
-              <Button size="small" onClick={deleteHandler}>
-                Delete
-              </Button>
-            </CardActions>
-          </>
+            <div>
+              <CardContent>
+                <Typography gutterBottom variant="h5" component="div">
+                  {post.title}
+                </Typography>
+              </CardContent>
+              <CardContent>
+                <Typography gutterBottom variant="p">
+                  {post.content}
+                </Typography>
+              </CardContent>
+              <CardActions sx={{ height: 50 }}>
+                <Button size="small" onClick={editHandler}>
+                  Edit
+                </Button>
+                <Button size="small" onClick={deleteHandler}>
+                  Delete
+                </Button>
+              </CardActions>
+            </div>
+          </div>
         </Card>
       )}
     </>
